remove UserApi namespace and replace it with UserService
