+++ src/api/users/user.ts
@@ -1,10 +1,27 @@
-import { Profile, User, UserSearchResults } from "@/types/user-types";
 import { ApiError } from "../api-error";
 import { endpoints } from "../endpoints";
 
-export namespace UserApi {
+export type User = {
+    id: string;
+    userName: string;
+    profilePicture: number;
+    createdOnUTC: Date;
+    followerCount: number;
+    followeeCount: number;
+}
 
-    export const Me = async (): Promise<User | null> => {
+export type Profile = {
+    user: User;
+    isFollowing: boolean;
+}
+
+export type UserSearchResults = {
+    items: { user: User; }[];
+}
+
+export const UserService = {
+
+    me: async (): Promise<User | null> => {
         try {    
             const response = await fetch(endpoints.users.me(), {
                 method: "GET",
@@ -40,9 +57,9 @@ export namespace UserApi {
             console.error("Caught exception fetching me: ", apiError);
             return null;
         }
-    }
+    },
 
-    export const User = async (id: string): Promise<User | null> => {
+    user: async (id: string): Promise<User | null> => {
         try {
             const response = await fetch(endpoints.users.getById(id), {
                 method: "GET",
@@ -76,9 +93,9 @@ export namespace UserApi {
             console.error("Caught exception fetching user: ", apiError);
             return null;
         }
-    }
+    },
 
-    export const Profile = async(id: string): Promise<Profile | null> => {
+    profile: async(id: string): Promise<Profile | null> => {
         try {
             const response = await fetch(endpoints.users.profile(id), {
                 method: "GET",
@@ -114,9 +131,9 @@ export namespace UserApi {
             console.error("Caught exception fetching user profile: ", apiError);
             return null;
         }
-    }
+    },
 
-    export const Search = async (query: string): Promise<UserSearchResults | null> => {
+    search: async (query: string): Promise<UserSearchResults | null> => {
         try {
             const response = await fetch(endpoints.users.search(query), {
                 method: "GET",
@@ -139,6 +156,6 @@ export namespace UserApi {
             console.error("UserSearch fetch failed:", error);
             return null;
         }
-    };
+    }
 
 }
